move config, basemodel and bigquery client out of main app.py
