+++ model/login.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+
+class LoginRequest(BaseModel):
+    username: str
+    password: str
+
+class LoginResponse(BaseModel):
+    access_token: str
+    token_type: str
+    user_id: str
+    username: str
+    email: str
+    role: str
+
